requirements.txt
asgiref==3.7.2
attrs==23.1.0
django~=4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
channels[daphne]~=4.0.0
channels_redis~=4.0.0
redis!=4.4.0
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.0
openai
Faker==21.0.0
django-seed==0.3.1